+++ tests/TestClient.py
@@ -3,6 +3,7 @@ import pytest
 import mock
 import faker
 import random
+import struct
 import sys
 from faker.providers import BaseProvider
 from thrift.transport.TSocket import TSocket
@@ -18,6 +19,20 @@ from milvus.client.Exceptions import (
 from milvus.thrift import ttypes, MilvusService
 
 LOGGER = logging.getLogger(__name__)
+
+
+def gen_one_binary(topk):
+    ids = [random.randrange(10000000, 99999999) for _ in range(topk)]
+    distances = [random.random() for _ in range(topk)]
+    return ttypes.TopKQueryBinResult(struct.pack(str(topk) + 'l', *ids), struct.pack(str(topk) + 'd', *distances))
+
+
+def gen_nq_binaries(nq, topk):
+    return [gen_one_binary(topk) for _ in range(nq)]
+
+
+def fake_query_bin_result(nq, topk):
+    return gen_nq_binaries(nq, topk)
 
 
 class FakerProvider(BaseProvider):
@@ -71,13 +86,14 @@ class TestConnection:
     param = {'host': 'localhost', 'port': '9898'}
 
     @mock.patch.object(TSocket, 'open')
-    def test_true_connect(self, open):
+    @mock.patch.object(Milvus, 'connected')
+    def test_true_connect(self, open, connected):
         open.return_value = None
+        connected.return_value = True
         cnn = Milvus()
 
         cnn.connect(**self.param)
         assert cnn.status.OK
-        assert cnn.connected
 
         _ = cnn.connect(**self.param)
         status = cnn.connect()
@@ -123,7 +139,7 @@ class TestConnection:
         cnn = Milvus()
         with pytest.raises(NotConnectError):
             cnn.connect(host='123.0.0.2', timeout=100)
-        assert not cnn.connected
+        assert not cnn.connected()
 
     @mock.patch.object(TSocket, 'open')
     def test_uri_runtime_error(self, open):
@@ -135,11 +151,13 @@ class TestConnection:
         cnn.connect()
         assert cnn.status.OK()
 
-    @mock.patch.object(TTransport.TBufferedTransport, 'close')
+    @mock.patch.object(TSocket, 'close')
+    @mock.patch.object(TSocket, 'isOpen')
     @mock.patch.object(TSocket, 'open')
-    def test_disconnected(self, close, open):
+    def test_disconnected(self, close, isOpen, open):
         close.return_value = None
         open.return_value = None
+        isOpen.return_value = True
 
         cnn = Milvus()
         cnn.connect(**self.param)
@@ -158,16 +176,18 @@ class TestTable:
     @pytest.fixture
     @mock.patch.object(TSocket, 'open')
     def client(self, open):
-        param = {'host': 'localhost', 'port': '5000'}
+        param = {'host': 'localhost', 'port': '5123'}
         open.return_value = None
 
         cnn = Milvus()
         cnn.connect(**param)
         return cnn
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'CreateTable')
-    def test_create_table(self, CreateTable, client):
+    def test_create_table(self, CreateTable, connected, client):
         CreateTable.return_value = None
+        connected.return_value = True
 
         param = table_schema_factory()
         param['table_name'] = None
@@ -222,9 +242,11 @@ class TestTable:
             res = client.create_table(param)
             assert res == Status.CONNECT_FAILED
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'DeleteTable')
-    def test_delete_table(self, DeleteTable, client):
+    def test_delete_table(self, DeleteTable, connected, client):
         DeleteTable.return_value = None
+        connected.return_value = True
         table_name = 'fake_table_name'
         res = client.delete_table(table_name)
         assert res.OK
@@ -236,30 +258,36 @@ class TestTable:
             LOGGER.info(res)
             assert res == Status.CONNECT_FAILED
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'HasTable')
-    def test_has_table(self, HasTable, client):
+    def test_has_table(self, HasTable, connected, client):
+        connected.return_value = True
         HasTable.return_value = True
+
         table_name = fake.table_name()
-        assert client.has_table(table_name)
+        result = client.has_table(table_name)
+        assert result
 
         HasTable.return_value = False
-        assert not client.has_table(table_name)
-
+        result = client.has_table(table_name)
+        assert not result
 
 class TestVector:
 
     @pytest.fixture
     @mock.patch.object(TSocket, 'open')
     def client(self, open):
-        param = {'host': 'localhost', 'port': '5000'}
+        param = {'host': 'localhost', 'port': '5123'}
         open.return_value = None
 
         cnn = Milvus()
         cnn.connect(**param)
         return cnn
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'AddVector')
-    def test_add_vector(self, AddVector, client):
+    def test_add_vector(self, AddVector, connected, client):
+        connected.return_value = True
         AddVector.return_value = ['a', 'a']
 
         param = {
@@ -287,25 +315,33 @@ class TestVector:
             res, ids = client.add_vectors(**param)
             assert res == Status.CONNECT_FAILED
 
-    @mock.patch.object(MilvusService.Client, 'SearchVector')
-    def test_search_vector(self, SearchVector, client):
-        SearchVector.return_value = [ttypes.TopKQueryResult([ttypes.QueryResult(111, 111)])]
+    @mock.patch.object(Milvus, 'connected')
+    @mock.patch.object(MilvusService.Client, 'SearchVector2')
+    def test_search_vector(self, SearchVector2, connected, client):
+        topk = random.randint(1, 10)
+        query_records = records_factory(256)
+        SearchVector2.return_value = fake_query_bin_result(1, topk)
+        connected.return_value = True
         param = {
             'table_name': fake.table_name(),
-            'query_records': records_factory(256),
-            'top_k': random.randint(1, 10)
+            'query_records': query_records,
+            'top_k': topk
         }
         res, results = client.search_vectors(**param)
         assert res.OK()
         assert isinstance(results, (list, TopKQueryResult))
 
-    @mock.patch.object(MilvusService.Client, 'SearchVector')
-    def test_search_vector_with_range(self, SearchVector, client):
-        SearchVector.return_value = [ttypes.TopKQueryResult([ttypes.QueryResult(111, 111)])]
+    @mock.patch.object(Milvus, 'connected')
+    @mock.patch.object(MilvusService.Client, 'SearchVector2')
+    def test_search_vector_with_range(self, SearchVector2, connected, client):
+        topk = random.randint(1, 10)
+        query_records = records_factory(256)
+        SearchVector2.return_value = fake_query_bin_result(1, topk)
+        connected.return_value = True
         param = {
             'table_name': fake.table_name(),
-            'query_records': records_factory(256),
-            'top_k': random.randint(1, 10),
+            'query_records': query_records,
+            'top_k': topk,
             'query_ranges': query_ranges_factory()
 
         }
@@ -313,7 +349,9 @@ class TestVector:
         assert res.OK()
         assert isinstance(results, (list, TopKQueryResult))
 
-    def test_false_vector(self, client):
+    @mock.patch.object(Milvus, 'connected')
+    def test_false_vector(self, connected, client):
+        connected.return_value = True
         param = {
             'table_name': fake.table_name(),
             'query_records': records_factory(256),
@@ -346,9 +384,12 @@ class TestVector:
         with pytest.raises(ParamError):
             res, results = client.search_vectors(**param)
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'SearchVectorInFiles')
-    def test_search_in_files(self, SearchVectorInFiles, client):
+    def test_search_in_files(self, SearchVectorInFiles, connected, client):
         SearchVectorInFiles.return_value = [ttypes.TopKQueryResult([ttypes.QueryResult(111, 111)])]
+        connected.return_value = True
+
         param = {
             'table_name': fake.table_name(),
             'query_records': records_factory(256),
@@ -369,9 +410,11 @@ class TestVector:
             sta, results = client.search_vectors_in_files(**param)
             assert sta == Status.CONNECT_FAILED
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'DescribeTable')
-    def test_describe_table(self, DescribeTable, client):
+    def test_describe_table(self, DescribeTable, connected, client):
         DescribeTable.return_value = ttypes.TableSchema(**table_schema_factory())
+        connected.return_value = True
 
         table_name = fake.table_name()
         res, table_schema = client.describe_table(table_name)
@@ -385,9 +428,11 @@ class TestVector:
             assert not res.OK()
             assert not table_schema
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'ShowTables')
-    def test_show_tables(self, ShowTables, client):
+    def test_show_tables(self, ShowTables, connected, client):
         ShowTables.return_value = [fake.table_name() for _ in range(10)]
+        connected.return_value = True
         res, tables = client.show_tables()
         assert res.OK()
         assert isinstance(tables, list)
@@ -398,9 +443,11 @@ class TestVector:
             assert not res.OK()
             assert not tables
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'GetTableRowCount')
-    def test_get_table_row_count(self, GetTableRowCount, client):
+    def test_get_table_row_count(self, GetTableRowCount, connected, client):
         GetTableRowCount.return_value = 22, None
+        connected.return_value = True
         res, count = client.get_table_row_count('fake_table')
         assert res.OK()
 
@@ -414,9 +461,11 @@ class TestVector:
         res = client.client_version()
         assert isinstance(res, str)
 
+    @mock.patch.object(Milvus, 'connected')
     @mock.patch.object(MilvusService.Client, 'Ping')
-    def test_server_status(self, Ping, client):
+    def test_server_status(self, Ping, connected, client):
         Ping.return_value = '00'
+        connected.return_value = True
         status, res = client.server_status()
         assert status.OK()
         assert res == 'OK'